Stopped a deep reading-compliance page from scanning the whole tenant
The latest-reading-per-operator LATERAL keys on created_by alone, which no index served — the closest, idx_*_flow_scheme_creator_date, leads on scheme_id. LIMIT hid that on page 0; OFFSET made it run once per operator, so ?page=999 timed out at the gateway. V41 adds (created_by, <time> DESC, id DESC), partial on deleted_at IS NULL, resolving the column per tenant vintage.

Counting first also skips the row query for a page starting past the last row — PageResponseDTO.of rejects it with a 400 regardless — and the offset is now long, so page * size no longer overflows to a negative OFFSET.

+++ backend/user-service/src/main/java/org/arghyam/jalsoochak/user/repository/PublicPumpOperatorRepository.java
@@ -809,7 +809,7 @@ public class PublicPumpOperatorRepository {
         }
     }
 
-    public List<PumpOperatorReadingComplianceRowDTO> listReadingCompliance(String schemaName, int offset, int limit) {
+    public List<PumpOperatorReadingComplianceRowDTO> listReadingCompliance(String schemaName, long offset, int limit) {
         validateSchemaName(schemaName);
         String timeColumn = resolveFlowReadingTimeColumn(schemaName);
         String confirmedExpr = resolveConfirmedReadingExpression(schemaName, "fr");
@@ -872,7 +872,7 @@ public class PublicPumpOperatorRepository {
             Long pumpOperatorId,
             LocalDate startDate,
             LocalDate endDate,
-            int offset,
+            long offset,
             int limit
     ) {
         validateSchemaName(schemaName);
